% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotrade-package.R
\docType{package}
\name{autotrade-package}
\alias{autotrade}
\alias{autotrade-package}
\title{autotrade: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
