% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_vctr.R
\name{backtest_vctr}
\alias{backtest_vctr}
\title{Backtest Positions in a Single Security Using Daily OHLCV Data}
\usage{
backtest_vctr(dt_ohlcvp, initial_cash = 1000L, in_place = FALSE)
}
\description{
Backtest Positions in a Single Security Using Daily OHLCV Data
}
