% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_drawdown.R
\name{max_drawdown}
\alias{max_drawdown}
\title{Calculate Maximum Drawdown Size within Current Drawdown}
\usage{
max_drawdown(x)
}
\arguments{
\item{x}{\verb{[numeric | xts]} A numeric vector or xts object containing price
data. Cannot contain missing values or multiple columns}
}
\value{
\verb{[class(x)]} An object of the same class as \code{x} with maximum drawdowns
at each value
}
\description{
Calculate Maximum Drawdown Size within Current Drawdown
}
