% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlib.R
\name{qlib}
\alias{qlib}
\title{Load Packages Quickly and Quietly}
\usage{
qlib(..., .warn_conflicts = FALSE, .quietly = TRUE)
}
\arguments{
\item{...}{\verb{[chr | sym]} One or more packages to attach; can be character
vectors or symbols}

\item{.warn_conflicts}{\verb{[lgl]} Should warnings be printed regarding conflicts
when attaching new packages?}

\item{.quietly}{\verb{[lgl]} Should all messages be suppressed?}
}
\value{
A \code{character} vector of attached packages (invisibly)
}
\description{
Load Packages Quickly and Quietly
}
